+++ locacao-service/src/main/java/com/example/locacao_service/repository/CarrinhoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface CarrinhoRepository extends JpaRepository<Carrinho, BigInteger> {
+public interface CarrinhoRepository extends JpaRepository<Carrinho, Long> {
 }
